Move tutorial views into their own file.
