+++ zproject/urls.py
@@ -124,8 +124,8 @@ urlpatterns += patterns('zerver.views',
     url(r'^json/fetch_api_key$',            'json_fetch_api_key'),
     url(r'^json/update_active_status$',     'json_update_active_status'),
     url(r'^json/get_active_statuses$',      'json_get_active_statuses'),
-    url(r'^json/tutorial_send_message$',    'json_tutorial_send_message'),
-    url(r'^json/tutorial_status$',          'json_tutorial_status'),
+    url(r'^json/tutorial_send_message$',    'tutorial.json_tutorial_send_message'),
+    url(r'^json/tutorial_status$',          'tutorial.json_tutorial_status'),
     url(r'^json/change_enter_sends$',       'user_settings.json_change_enter_sends'),
     url(r'^json/get_profile$',              'json_get_profile'),
     url(r'^json/report_error$',             'report.json_report_error'),
